chore: #11 context 사용하는 곳으로 이동, 파일 제거
