--- src/store/context/CheckBoxContext.ts
@@ -1,8 +1,0 @@
-import { createContext } from "react";
-
-export type CheckBoxContextState = {
-  isChecked: (value: string) => boolean;
-  onChange: (value: string) => void;
-};
-
-export const CheckBoxContext = createContext<CheckBoxContextState | undefined>(undefined);
